Refactor event-watcher and move code to util
